refactor: Service 네이밍 수정
