+++ data/src/main/java/com/susu/data/network/TokenAuthenticator.kt
@@ -1,6 +1,7 @@
 package com.susu.data.network
 
 import com.susu.data.model.request.RefreshTokenRequest
+import com.susu.data.network.service.TokenService
 import com.susu.domain.repository.TokenRepository
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.runBlocking
@@ -12,7 +13,7 @@ import javax.inject.Inject
 
 class TokenAuthenticator @Inject constructor(
     private val tokenRepository: TokenRepository,
-    private val authService: AuthService,
+    private val tokenService: TokenService,
 ) : Authenticator {
     override fun authenticate(route: Route?, response: Response): Request? {
         // 1. refresh token으로 갱신 요청
@@ -25,7 +26,7 @@ class TokenAuthenticator @Inject constructor(
 
         return runBlocking {
             // 2. access token 갱신
-            val tokenResponse = authService.refreshAccessToken(RefreshTokenRequest(refreshToken))
+            val tokenResponse = tokenService.refreshAccessToken(RefreshTokenRequest(refreshToken))
 
             // 2-1. 정상적으로 받지 못하면 request token 까지 만료된 것.
             if (tokenResponse.isSuccessful.not() || tokenResponse.body() == null) {
