+++ data/src/main/java/com/susu/data/network/service/SignUpService.kt
@@ -1,23 +1,15 @@
-package com.susu.data.network
+package com.susu.data.network.service
 
 import com.susu.data.model.TokenEntity
 import com.susu.data.model.UserEntity
-import com.susu.data.model.request.AccessTokenRequest
 import com.susu.data.model.response.ValidRegisterResponse
-import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
 import retrofit2.http.Path
 import retrofit2.http.Query
 
-interface ApiService {
-    @POST("oauth/{provider}/login")
-    suspend fun login(
-        @Path("provider") provider: String,
-        @Body accessTokenRequest: AccessTokenRequest,
-    ): TokenEntity
-
+interface SignUpService {
     @POST("oauth/{provider}/sign-up")
     suspend fun signUp(
         @Path("provider") provider: String,
@@ -30,10 +22,4 @@ interface ApiService {
         @Path("provider") provider: String,
         @Query("accessToken") accessToken: String,
     ): ValidRegisterResponse
-
-    @POST("auth/logout")
-    suspend fun logout(): Response<Unit?>
-
-    @POST("auth/withdraw")
-    suspend fun withdraw(): Response<Unit?>
 }
