+++ data/src/main/java/com/susu/data/network/service/TokenService.kt
@@ -1,4 +1,4 @@
-package com.susu.data.network
+package com.susu.data.network.service
 
 import com.susu.data.model.TokenEntity
 import com.susu.data.model.request.RefreshTokenRequest
@@ -6,7 +6,7 @@ import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
 
-interface AuthService {
+interface TokenService {
     @POST("auth/token/refresh")
     suspend fun refreshAccessToken(
         @Body refreshTokenRequest: RefreshTokenRequest,
