+++ data/src/main/java/com/susu/data/network/service/UserService.kt
@@ -1,0 +1,22 @@
+package com.susu.data.network.service
+
+import com.susu.data.model.TokenEntity
+import com.susu.data.model.request.AccessTokenRequest
+import retrofit2.Response
+import retrofit2.http.Body
+import retrofit2.http.POST
+import retrofit2.http.Path
+
+interface UserService {
+    @POST("oauth/{provider}/login")
+    suspend fun login(
+        @Path("provider") provider: String,
+        @Body accessTokenRequest: AccessTokenRequest,
+    ): TokenEntity
+
+    @POST("auth/logout")
+    suspend fun logout(): Response<Unit?>
+
+    @POST("auth/withdraw")
+    suspend fun withdraw(): Response<Unit?>
+}
