+++ data/src/main/java/com/susu/data/repository/AuthRepositoryImpl.kt
@@ -5,17 +5,19 @@ import com.susu.data.model.SnsProviders
 import com.susu.data.model.request.AccessTokenRequest
 import com.susu.data.model.toData
 import com.susu.data.model.toDomain
-import com.susu.data.network.ApiService
+import com.susu.data.network.service.SignUpService
+import com.susu.data.network.service.UserService
 import com.susu.domain.repository.AuthRepository
 import javax.inject.Inject
 
 class AuthRepositoryImpl @Inject constructor(
-    private val apiService: ApiService,
+    private val userService: UserService,
+    private val signUpService: SignUpService,
 ) : AuthRepository {
     override suspend fun login(
         accessToken: String,
     ) = kotlin.runCatching {
-        apiService.login(
+        userService.login(
             provider = SnsProviders.Kakao.path,
             accessTokenRequest = AccessTokenRequest(accessToken),
         ).toDomain()
@@ -25,7 +27,7 @@ class AuthRepositoryImpl @Inject constructor(
         snsAccessToken: String,
         user: User,
     ) = runCatching {
-        apiService.signUp(
+        signUpService.signUp(
             provider = SnsProviders.Kakao.path,
             accessToken = snsAccessToken,
             user = user.toData(),
@@ -35,17 +37,17 @@ class AuthRepositoryImpl @Inject constructor(
     override suspend fun canRegister(
         accessToken: String,
     ): Boolean {
-        return apiService.checkValidRegister(
+        return signUpService.checkValidRegister(
             provider = SnsProviders.Kakao.path,
             accessToken = accessToken,
         ).canRegister
     }
 
     override suspend fun logout() {
-        apiService.logout()
+        userService.logout()
     }
 
     override suspend fun withdraw() {
-        apiService.withdraw()
+        userService.withdraw()
     }
 }
